Rearrange convolution output if operands are not swapped in the weight update pass
Summary:
The rearrangement is done using a transpose. The maximum number
of input rows in the rearrangement is capped at 16.

The types of the columns and rows are changed to unsigned short
from unsigned integers to reduce state

The memory saving are very small for Resnet50 batch size 4
but compute cost in IPUModel improves by 2.2%

Test Plan: ./test.sh poplibs

Reviewers: richard, tomm

Reviewed By: tomm

Subscribers: samk, timp

Maniphest Tasks: T5315

Differential Revision: https://phabricator.sourcevertex.net/D8052

+++ lib/poplin/codelets/poplinCodelets.cpp
@@ -838,8 +838,8 @@ public:
   Vector<Output<Vector<T, ONE_PTR,8>>, ONE_PTR> dst;
   // TODO specialize the vertex based on the value of this field to avoid extra
   // memory usage.
-  unsigned numSrcRows;
-  unsigned numSrcColumns;
+  unsigned short numSrcRows;
+  unsigned short numSrcColumns;
 
   IS_EXTERNAL_CODELET(true);
